doc: apply markdownlint suggestions
I added empty lines before code fences and alt text for images. You might
also notice that I removed a trailing question mark in the "Why rash" title.
That's also a markdownlint recommendation, and I think it makes sense. I think
an affirmative title is more compelling that a one with a question.

Signed-off-by: Miguel González <migueljgonz@gmail.com>